Small changes for translation mechanism (#12)
* Add basic app routes without middleware

* Small changes eg. title, translation, inertia

* Delegate translation mechanism to useTranslation.js
