+++ resources/js/Components/LanguageSwitcher.vue
@@ -4,24 +4,21 @@
       v-for="(code, name) in availableLocales"
       :key="code"
       href="#"
-      @click.prevent="switchLanguage(code)"
+      @click.prevent="onLanguageChange(code)"
       class="px-2 py-1 rounded transition-colors duration-200"
       :class="currentLocale === code ? 'bg-indigo-100 text-indigo-800 font-medium' : 'text-gray-600 hover:text-indigo-600'"
     >
-      {{ name }}
+      {{ trans(name) }}
     </a>
   </div>
 </template>
 
 <script setup>
-import { usePage } from '@inertiajs/vue3';
-import { computed } from 'vue';
+import { useTranslation } from '@/Composables/useTranslation';
+const { trans } = useTranslation();
+const { currentLocale, availableLocales, switchLanguage } = useTranslation();
 
-const page = usePage();
-const currentLocale = computed(() => page.props.locale || 'en');
-const availableLocales = computed(() => page.props.availableLocales || {'English': 'en', 'Polish': 'pl'});
-
-function switchLanguage(locale) {
-  window.location.href = `/language/${locale}`;
+function onLanguageChange(code) {
+  switchLanguage(code);
 }
 </script>
