+++ resources/js/app.ts
@@ -2,8 +2,9 @@ import '../css/app.css'
 import { createApp, h, type DefineComponent } from 'vue'
 import { createInertiaApp } from '@inertiajs/vue3'
 import { resolvePageComponent } from 'laravel-vite-plugin/inertia-helpers'
+import LanguageSwitcher from "@/Components/LanguageSwitcher.vue";
 
-const appName = import.meta.env.VITE_APP_NAME || 'Laravel'
+const appName = import.meta.env.VITE_APP_NAME || 'Questify'
 
 createInertiaApp({
   title: (title) => `${title} - ${appName}`,
@@ -11,6 +12,7 @@ createInertiaApp({
   setup({ el, App, props, plugin }) {
     createApp({ render: () => h(App, props) })
       .use(plugin)
+      .component('LanguageSwitcher', LanguageSwitcher)
       .mount(el)
   },
   progress: {
