+++ resources/lang/en.json
@@ -1,5 +1,7 @@
 {
     "Start adventure!": "Start adventure!",
     "Begin Your Quest": "Begin Your Quest",
-    "Transform your productivity into an epic adventure!": "Transform your productivity into an epic adventure!"
+    "Transform your productivity into an epic adventure!": "Transform your productivity into an epic adventure!",
+    "English": "English",
+    "Polish": "Polish"
 }
