+++ resources/lang/pl.json
@@ -1,5 +1,7 @@
 {
     "Start adventure!": "Rozpocznij przygodę!",
     "Begin Your Quest": "Rozpocznij swoją misję",
-    "Transform your productivity into an epic adventure!": "Przekształć swoją produktywność w epicką przygodę!"
+    "Transform your productivity into an epic adventure!": "Przekształć swoją produktywność w epicką przygodę!",
+    "English": "Angielski",
+    "Polish": "Polski"
 }
