+++ resources/views/app.blade.php
@@ -3,7 +3,7 @@
 <head>
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1">
-    <title>Laravel + Inertia</title>
+    <title>Questify</title>
     @vite(['resources/css/app.css', 'resources/js/app.ts'])
     @inertiaHead
 </head>
